+++ base-plugin/src/main/java/com/ss/android/ugc/bytex/base/ByteXTransform.java
@@ -1,10 +1,12 @@
 package com.ss.android.ugc.bytex.base;
 
+import com.android.build.api.transform.TransformException;
+import com.android.build.api.transform.TransformInvocation;
 import com.ss.android.ugc.bytex.common.CommonTransform;
 import com.ss.android.ugc.bytex.common.IPlugin;
-import com.ss.android.ugc.bytex.common.log.LevelLog;
 import com.ss.android.ugc.bytex.transformer.TransformContext;
 
+import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 
@@ -32,6 +34,12 @@ public class ByteXTransform extends CommonTransform<Context> {
     @Override
     public boolean shouldSaveCache() {
         return context.extension.isShouldSaveCache() && super.shouldSaveCache();
+    }
+
+    @Override
+    protected void afterTransform(TransformInvocation transformInvocation) throws TransformException, InterruptedException, IOException {
+        super.afterTransform(transformInvocation);
+        context.releaseContext();
     }
 
     @Override