+++ common/src/main/java/com/ss/android/ugc/bytex/common/graph/cache/IClassCache.kt
@@ -1,0 +1,9 @@
+package com.ss.android.ugc.bytex.common.graph.cache
+
+import com.ss.android.ugc.bytex.common.graph.ClassEntity
+import java.io.File
+
+/**
+ * Created by yangzhiqian on 2020/7/13<br/>
+ */
+interface IClassCache : ICache<File, List<ClassEntity>>