+++ refer-check-plugin/src/main/java/com/ss/android/ugc/bytex/refercheck/ReferCheckContext.java
@@ -182,11 +182,15 @@ public class ReferCheckContext extends BaseContext<ReferCheckExtension> {
 
     public void releaseContext() {
         super.releaseContext();
-        methodCache.release();
-        fieldCache.release();
+        if (methodCache != null) {
+            methodCache.release();
+            methodCache = null;
+        }
+        if (fieldCache != null) {
+            fieldCache.release();
+            fieldCache = null;
+        }
         notAccessMethods.clear();
-        methodCache = null;
-        fieldCache = null;
     }
 
 }